Fix: Resolve database schema mismatches between init.sql and TypeORM entities
- Fix source_id type mismatch: Changed from BIGINT to INTEGER in init.sql and entities
  - sources.id uses INTEGER (OID values are small: 001-448)
  - news_articles.source_id must match as INTEGER for proper foreign key reference
- Update Source entity: Use @Column primary instead of @PrimaryGeneratedColumn for manual OID assignment
- Update MarketSummary entity and service to match init.sql schema
  - Changed from symbol/price/market/currency to marketType/currentValue/changeValue/volume/tradingValue/highValue/lowValue
  - Updated marketSummaryService to use new column names
- Ensure all entities match database schema defined in init.sql

+++ backend/api/src/entities/NewsArticle.ts
@@ -28,7 +28,7 @@ export class NewsArticle {
     @Column({ type: 'varchar', length: 1000, nullable: true, name: 'image_url' })
     imageUrl?: string;
 
-    @Column({ type: 'bigint', name: 'source_id' })
+    @Column({ type: 'integer', name: 'source_id' })
     sourceId: number;
 
     @Column({ type: 'bigint', name: 'category_id' })
